+++ config/internals/envVars.js
@@ -1,33 +1,48 @@
+/**
+ * Helper for resolving environment specific configuration.
+ *
+ * TODO: Detailed documentation.
+ */
+
+import appRootDir from 'app-root-dir';
+import colors from 'colors/safe';
 import dotenv from 'dotenv';
 import fs from 'fs';
 import path from 'path';
-import appRootDir from 'app-root-dir';
 import userHome from 'user-home';
-import colors from 'colors/safe';
+
 import pkg from '../../package.json';
 
+// PRIVATES
+
+// :: (Boolean, () => Element) -> ?Any
+const onlyIf = (a, b) => (a ? b() : null);
+
+// :: (Boolean)
+const removeEmpty = a => a.filter(x => x != null);
+
 function registerEnvFile() {
-  const envName = process.env.NODE_ENV || 'development';
+  const CONF_ENV = process.env.CONF_ENV;
   const envFile = '.env';
 
   // This is the order in which we will try to resolve an environment configuration
   // file.
-  const envFileResolutionOrder = [
+  const envFileResolutionOrder = removeEmpty([
     // Is there an environment config file at the app root for our target
     // environment name?
     // e.g. /projects/react-universally/.env.development
-    path.resolve(appRootDir.get(), `${envFile}.${envName}`),
+    onlyIf(CONF_ENV, () => path.resolve(appRootDir.get(), `${envFile}.${CONF_ENV}`)),
     // Is there an environment config file at the app root?
     // e.g. /projects/react-universally/.env
     path.resolve(appRootDir.get(), envFile),
     // Is there an environment config file in the executing user's home dir
     // that is targetting the specific environment?
     // e.g. /Users/ctrlplusb/.config/react-universally/.env.development
-    path.resolve(userHome, '.config', pkg.name, `${envFile}.${envName}`),
+    onlyIf(CONF_ENV, () => path.resolve(userHome, '.config', pkg.name, `${envFile}.${CONF_ENV}`)),
     // Is there an environment config file in the executing user's home dir?
     // e.g. /Users/ctrlplusb/.config/react-universally/.env
     path.resolve(userHome, '.config', pkg.name, envFile),
-  ];
+  ]);
 
   // Find the first env file path match.
   const envFilePath = envFileResolutionOrder.find(filePath => fs.existsSync(filePath));
@@ -45,18 +60,20 @@ function registerEnvFile() {
 // env file.
 registerEnvFile();
 
-export function getStringEnvVar(name, defaultVal) {
+// EXPORTED HELPERS
+
+export function string(name, defaultVal) {
   return process.env[name] || defaultVal;
 }
 
-export function getIntEnvVar(name, defaultVal) {
+export function int(name, defaultVal) {
   return process.env[name]
     ? parseInt(process.env[name], 10)
     : defaultVal;
 }
 
-export function getBoolVar(name, defaultVal) {
+export function bool(name, defaultVal) {
   return process.env[name]
-    ? process.env[name] === 'true'
+    ? process.env[name] === 'true' || process.env[name] === '1'
     : defaultVal;
 }