+++ src/server/middleware/reactApplication/index.js
@@ -1,16 +1,17 @@
 
 import React from 'react';
-import { renderToString } from 'react-dom/server';
+import { renderToString, renderToStaticMarkup } from 'react-dom/server';
 import { ServerRouter, createServerRenderContext } from 'react-router';
-import { CodeSplitProvider, createRenderContext } from 'code-split-component';
+import { withAsyncComponents } from 'react-async-component';
 import Helmet from 'react-helmet';
-import generateHTML from './generateHTML';
-import DemoApp from '../../../shared/components/DemoApp';
+
 import config from '../../../../config';
 
+import ServerHTML from './ServerHTML';
+import DemoApp from '../../../shared/components/DemoApp';
+
 /**
- * An express middleware that is capabable of service our React application,
- * supporting server side rendering of the application.
+ * React application middleware, supports server side rendering.
  */
 function reactApplicationMiddleware(request, response) {
   // We should have had a nonce provided to us.  See the server/index.js for
@@ -29,10 +30,7 @@ function reactApplicationMiddleware(request, response) {
     }
     // SSR is disabled so we will just return an empty html page and will
     // rely on the client to initialize and render the react application.
-    const html = generateHTML({
-      // Nonce which allows us to safely declare inline scripts.
-      nonce,
-    });
+    const html = renderToStaticMarkup(<ServerHTML nonce={nonce} />);
     response.status(200).send(html);
     return;
   }
@@ -41,56 +39,51 @@ function reactApplicationMiddleware(request, response) {
   // query for the results of the render.
   const reactRouterContext = createServerRenderContext();
 
-  // We also create a context for our <CodeSplitProvider> which will allow us
-  // to query which chunks/modules were used during the render process.
-  const codeSplitContext = createRenderContext();
-
-  // Create our React application and render it into a string.
-  const reactAppString = renderToString(
-    <CodeSplitProvider context={codeSplitContext}>
-      <ServerRouter location={request.url} context={reactRouterContext}>
-        <DemoApp />
-      </ServerRouter>
-    </CodeSplitProvider>,
+  // Create our React application.
+  const app = (
+    <ServerRouter location={request.url} context={reactRouterContext}>
+      <DemoApp />
+    </ServerRouter>
   );
 
-  // Generate the html response.
-  const html = generateHTML({
-    // Provide the full app react element.
-    reactAppString,
-    // Nonce which allows us to safely declare inline scripts.
-    nonce,
-    // Running this gets all the helmet properties (e.g. headers/scripts/title etc)
-    // that need to be included within our html.  It's based on the rendered app.
-    // @see https://github.com/nfl/react-helmet
-    helmet: Helmet.rewind(),
-    // We provide our code split state so that it can be included within the
-    // html, and then the client bundle can use this data to know which chunks/
-    // modules need to be rehydrated prior to the application being rendered.
-    codeSplitState: codeSplitContext.getState(),
+  // Wrap our app with react-async-component helper so that our async components
+  // will be resolved and rendered with the response.
+  withAsyncComponents(app).then(({ appWithAsyncComponents, state, STATE_IDENTIFIER }) => {
+    // Render the app to a string.
+    const reactAppString = renderToString(appWithAsyncComponents);
+
+    // Generate the html response.
+    const html = renderToStaticMarkup(
+      <ServerHTML
+        reactAppString={reactAppString}
+        nonce={nonce}
+        helmet={Helmet.rewind()}
+        asyncComponents={{ state, STATE_IDENTIFIER }}
+      />,
+    );
+
+    // Get the render result from the server render context.
+    const renderResult = reactRouterContext.getResult();
+
+    // Check if the render result contains a redirect, if so we need to set
+    // the specific status and redirect header and end the response.
+    if (renderResult.redirect) {
+      response.status(301).setHeader('Location', renderResult.redirect.pathname);
+      response.end();
+      return;
+    }
+
+    response
+      .status(
+        renderResult.missed
+          // If the renderResult contains a "missed" match then we set a 404 code.
+          // Our App component will handle the rendering of an Error404 view.
+          ? 404
+          // Otherwise everything is all good and we send a 200 OK status.
+          : 200,
+      )
+      .send(html);
   });
-
-  // Get the render result from the server render context.
-  const renderResult = reactRouterContext.getResult();
-
-  // Check if the render result contains a redirect, if so we need to set
-  // the specific status and redirect header and end the response.
-  if (renderResult.redirect) {
-    response.status(301).setHeader('Location', renderResult.redirect.pathname);
-    response.end();
-    return;
-  }
-
-  response
-    .status(
-      renderResult.missed
-        // If the renderResult contains a "missed" match then we set a 404 code.
-        // Our App component will handle the rendering of an Error404 view.
-        ? 404
-        // Otherwise everything is all good and we send a 200 OK status.
-        : 200,
-    )
-    .send(html);
 }
 
-export default (reactApplicationMiddleware);
+export default reactApplicationMiddleware;