+++ src/shared/components/DemoApp/AsyncAbout/index.js
@@ -1,0 +1,7 @@
+import { createAsyncComponent } from 'react-async-component';
+
+export default createAsyncComponent({
+  resolve: () => System.import('./About'),
+  ssrMode: 'boundary',
+  name: 'AsyncAbout',
+});