+++ src/shared/components/DemoApp/AsyncHome/index.js
@@ -1,0 +1,7 @@
+import { createAsyncComponent } from 'react-async-component';
+
+export default createAsyncComponent({
+  resolve: () => System.import('./Home'),
+  ssrMode: 'boundary',
+  name: 'AsyncHome',
+});