+++ tools/webpack/client.config.babel.js
@@ -1,6 +1,0 @@
-import webpackConfigFactory from './configFactory';
-
-export default function clientConfigFactory(options = {}) {
-  const { mode = 'development' } = options;
-  return webpackConfigFactory({ target: 'client', mode });
-}