+++ tools/webpack/server.config.babel.js
@@ -1,6 +1,0 @@
-import webpackConfigFactory from './configFactory';
-
-export default function serverConfigFactory(options = {}) {
-  const { mode = 'development' } = options;
-  return webpackConfigFactory({ target: 'server', mode });
-}